requirements.txt
matplotlib==3.9.1.post1
numpy==2.0.1
pandas==2.2.2
python_can==4.4.2
seaborn==0.13.2
torch==2.3.1+cu121
